Update the `HPX_GAS_BLOCK_BYTES_MAX` to be a GAS-specific value, and extend the

GAS interface to support `size_t` block sizes.

The HPX_GAS_BLOCK_BYTES_MAX can be used by the application to decompose data
structures in an application specific way, but it depends on the GAS that has
been selected dynamically. For example, our PGAS implementation limits blocks to
2^32 bytes, but the SMP implementation has no obvious bound.

In order to use this effectively, the block-size interface needs to be updated
everywhere to use `size_t` block sizes as well. Specific implementations that
need smaller block sizes can test dynamically and then do a narrowing cast.

See merge request !88

+++ include/hpx/cxx/global_ptr.h
@@ -138,13 +138,13 @@ class global_ptr {
   ///
   /// @param       addr The hpx_addr_t.
   /// @param          n The number of elements per block.
-  global_ptr(hpx_addr_t addr, uint32_t n) noexcept : _gbl_ptr(addr),
-                                                     _elems_per_blk(n) {
+  global_ptr(hpx_addr_t addr, size_t n) noexcept : _gbl_ptr(addr),
+    _elems_per_blk(n) {
   }
 
   /// A typed global pointer can be constructed from a global_ptr<void>, as long
   /// as the user provides a block size.
-  explicit global_ptr(const global_ptr<void>& gva, uint32_t n) noexcept
+  explicit global_ptr(const global_ptr<void>& gva, size_t n) noexcept
     : _gbl_ptr(gva.get()), _elems_per_blk(n) {
   }
 
@@ -159,7 +159,7 @@ class global_ptr {
   }
 
   /// Return the block size that this smart pointer encapsulates.
-  uint32_t get_block_size() const noexcept {
+  size_t get_block_size() const noexcept {
     return _elems_per_blk;
   }
