+++ libhpx/gas/agas/agas.h
@@ -70,18 +70,18 @@ int agas_memcpy(void *gas, hpx_addr_t to, hpx_addr_t from, size_t size,
 
 int agas_memcpy_sync(void *gas, hpx_addr_t to, hpx_addr_t from, size_t size);
 
-gva_t agas_lva_to_gva(agas_t *gas, void *lva, uint32_t bsize);
+gva_t agas_lva_to_gva(agas_t *gas, void *lva, size_t bsize);
 
-hpx_addr_t agas_alloc_local(size_t n, uint32_t bsize, uint32_t boundary,
+hpx_addr_t agas_alloc_local(size_t n, size_t bsize, uint32_t boundary,
                             uint32_t attr);
 
-hpx_addr_t agas_calloc_local(size_t n, uint32_t bsize, uint32_t boundary,
+hpx_addr_t agas_calloc_local(size_t n, size_t bsize, uint32_t boundary,
                              uint32_t attr);
 
-int64_t agas_sub_local(const agas_t *agas, gva_t lhs, gva_t rhs, uint32_t bsize);
+int64_t agas_sub_local(const agas_t *agas, gva_t lhs, gva_t rhs, size_t bsize);
 
 hpx_addr_t agas_add_local(const agas_t *agas, gva_t gva, int64_t bytes,
-                          uint32_t bsize);
+                          size_t bsize);
 
 void agas_free(void *gas, hpx_addr_t addr, hpx_addr_t rsync);
